refactored aslinop into own file
