+++ tests/test_linops/test_linearoperators.py
@@ -2,6 +2,7 @@
 import itertools
 import unittest
 
+import linops
 import numpy as np
 import scipy.sparse
